refactor(user): Изменения в части работы с таблицей БД

+++ src/Domain/Exceptions/YagoNotFoundException.cs
@@ -2,7 +2,7 @@
 {
     public class YagoNotFoundException : YagoException
     {
-        public YagoNotFoundException(string type, int id)
+        public YagoNotFoundException(string type, long id)
             : base(string.Format("Не удалось найти объект {0} с Id={1}.", type, id))
         { }
     }
